# src/main/resources/static/openapi.yaml
openapi: 3.0.0
info:
  description: API for managing event locations
  title: Event Management Service
  version: 1.0.0
servers:
- url: http://localhost:8081
security:
- bearerAuth: []
...
paths:
...
  /events:
    post:
      description: "Учтите валидацию всех полей. Все поля обязательные при создани\
        и, также должна быть соблюдена логика (например время не может быть в прош\
        лом, стомость > 0 и т.д.) Локация должна существовать и ее вместимость дол\
        жна позволять провести это мероприятие, т.е. кол-во мест должно быть доста\
        точно для всех участников. Организатор (создатель) мероприятия не учитывае\
        тся как участник."
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventCreateRequestDto'
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventDto'
          description: Мероприятие успешно создано. Возвращается созданное событие.
        "400":
          content:
            application/json:
              example:
                message: Некорректный запрос
                detailedMessage: Более детальное сообщение об ошибке
                dateTime: 2024-12-03T10:15:30
              schema:
                $ref: '#/components/schemas/ErrorMessageResponse'
          description: "Запрос с невалидными данными. Например, попытка занять зан\
            ятый логин или некорректные параметры запроса."
        "401":
          content:
            application/json:
              example:
                message: Необходима аутентификация
                detailedMessage: Более детальное сообщение об ошибке
                dateTime: 2024-12-03T10:15:30
              schema:
                $ref: '#/components/schemas/ErrorMessageResponse'
          description: Аутентификация отсутствует или не удалась
        "403":
          content:
            application/json:
              example:
                message: Недостаточно прав для выполнения операции
                detailedMessage: Более детальное сообщение об ошибке
                dateTime: 2024-12-03T10:15:30
              schema:
                $ref: '#/components/schemas/ErrorMessageResponse'
          description: У пользователя недостаточно прав для выполнения этой операции
        "404":
          content:
            application/json:
              example:
                message: Сущность не найдена
                detailedMessage: Более детальное сообщение об ошибке
                dateTime: 2024-12-03T10:15:30
              schema:
                $ref: '#/components/schemas/ErrorMessageResponse'
          description: Сущность не найдена.
        "500":
          content:
            application/json:
              example:
                message: Внутренняя ошибка сервера
                detailedMessage: Более детальное сообщение об ошибке
                dateTime: 2024-12-03T10:15:30
              schema:
                $ref: '#/components/schemas/ErrorMessageResponse'
          description: Внутренняя ошибка сервера
      summary: "Создание нового мероприятия. Allowed roles=[USER]"
      tags:
      - Events
  /events/{eventId}:
    delete:
      description: "Доступно только ADMIN, либо создателю мероприятия. Т.е. роль и\
        з JWT токена должна быть ADMIN, либо userId должен быть равен создателю ме\
        ропрития. Должно быть реализовано \"мягкое(soft) удаление\" - на самом дел\
        е строка в БД не удалется, только меняется статус мероприятия на CANCELLED.\
        \ При этом нужно учесть то, что не каждое мероприятие можно отменить (можн\
        о отменить только те мероприятия, которые еще не начались)."
      parameters:
      - description: Уникальный идентификатор мероприятия
        in: path
        name: eventId
        required: true
        schema:
          example: 10
          format: int64
          type: integer
      responses:
        "204":
          description: Мероприятие успешно удалено
        "400":
          content:
            application/json:
              example:
                message: Некорректный запрос
                detailedMessage: Более детальное сообщение об ошибке
                dateTime: 2024-12-03T10:15:30
              schema:
                $ref: '#/components/schemas/ErrorMessageResponse'
          description: "Запрос с невалидными данными. Например, попытка занять зан\
            ятый логин или некорректные параметры запроса."
        "401":
          content:
            application/json:
              example:
                message: Необходима аутентификация
                detailedMessage: Более детальное сообщение об ошибке
                dateTime: 2024-12-03T10:15:30
              schema:
                $ref: '#/components/schemas/ErrorMessageResponse'
          description: Аутентификация отсутствует или не удалась
        "403":
          content:
            application/json:
              example:
                message: Недостаточно прав для выполнения операции
                detailedMessage: Более детальное сообщение об ошибке
                dateTime: 2024-12-03T10:15:30
              schema:
                $ref: '#/components/schemas/ErrorMessageResponse'
          description: У пользователя недостаточно прав для выполнения этой операции
        "404":
          content:
            application/json:
              example:
                message: Сущность не найдена
                detailedMessage: Более детальное сообщение об ошибке
                dateTime: 2024-12-03T10:15:30
              schema:
                $ref: '#/components/schemas/ErrorMessageResponse'
          description: Сущность не найдена.
        "500":
          content:
            application/json:
              example:
                message: Внутренняя ошибка сервера
                detailedMessage: Более детальное сообщение об ошибке
                dateTime: 2024-12-03T10:15:30
              schema:
                $ref: '#/components/schemas/ErrorMessageResponse'
          description: Внутренняя ошибка сервера
      summary: "Удаление мероприятия. Allowed roles=[USER, ADMIN]"
      tags:
      - Events
    get:
      description: "Предусмотрите обработку ошибок, если мероприятие не найдено."
      parameters:
      - description: Уникальный идентификатор мероприятия
        in: path
        name: eventId
        required: true
        schema:
          example: 10
          format: int64
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventDto'
          description: Данные мероприятия
        "400":
          content:
            application/json:
              example:
                message: Некорректный запрос
                detailedMessage: Более детальное сообщение об ошибке
                dateTime: 2024-12-03T10:15:30
              schema:
                $ref: '#/components/schemas/ErrorMessageResponse'
          description: "Запрос с невалидными данными. Например, попытка занять зан\
            ятый логин или некорректные параметры запроса."
        "401":
          content:
            application/json:
              example:
                message: Необходима аутентификация
                detailedMessage: Более детальное сообщение об ошибке
                dateTime: 2024-12-03T10:15:30
              schema:
                $ref: '#/components/schemas/ErrorMessageResponse'
          description: Аутентификация отсутствует или не удалась
        "403":
          content:
            application/json:
              example:
                message: Недостаточно прав для выполнения операции
                detailedMessage: Более детальное сообщение об ошибке
                dateTime: 2024-12-03T10:15:30
              schema:
                $ref: '#/components/schemas/ErrorMessageResponse'
          description: У пользователя недостаточно прав для выполнения этой операции
        "404":
          content:
            application/json:
              example:
                message: Сущность не найдена
                detailedMessage: Более детальное сообщение об ошибке
                dateTime: 2024-12-03T10:15:30
              schema:
                $ref: '#/components/schemas/ErrorMessageResponse'
          description: Сущность не найдена.
        "500":
          content:
            application/json:
              example:
                message: Внутренняя ошибка сервера
                detailedMessage: Более детальное сообщение об ошибке
                dateTime: 2024-12-03T10:15:30
              schema:
                $ref: '#/components/schemas/ErrorMessageResponse'
          description: Внутренняя ошибка сервера
      summary: "Получение мероприятия по ID. Allowed roles=[USER, ADMIN]"
      tags:
      - Events
    put:
      description: "Доступно только ADMIN, либо создателю мероприятия. Т.е. роль и\
        з JWT токена должна быть ADMIN, либо userId должен быть равен создателю ме\
        ропрития. Учтите то, как можно менять мероприятие. Валидируйте входные зна\
        чения, например maxPlaces должно быть больше, чем уже записанных пользоват\
        елей, стоимость > 0, длительность > 0 и т.д."
      parameters:
      - description: Уникальный идентификатор мероприятия
        in: path
        name: eventId
        required: true
        schema:
          example: 10
          format: int64
          type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventUpdateRequestDto'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventDto'
          description: Обновленное мероприятие
        "400":
          content:
            application/json:
              example:
                message: Некорректный запрос
                detailedMessage: Более детальное сообщение об ошибке
                dateTime: 2024-12-03T10:15:30
              schema:
                $ref: '#/components/schemas/ErrorMessageResponse'
          description: "Запрос с невалидными данными. Например, попытка занять зан\
            ятый логин или некорректные параметры запроса."
        "401":
          content:
            application/json:
              example:
                message: Необходима аутентификация
                detailedMessage: Более детальное сообщение об ошибке
                dateTime: 2024-12-03T10:15:30
              schema:
                $ref: '#/components/schemas/ErrorMessageResponse'
          description: Аутентификация отсутствует или не удалась
        "403":
          content:
            application/json:
              example:
                message: Недостаточно прав для выполнения операции
                detailedMessage: Более детальное сообщение об ошибке
                dateTime: 2024-12-03T10:15:30
              schema:
                $ref: '#/components/schemas/ErrorMessageResponse'
          description: У пользователя недостаточно прав для выполнения этой операции
        "404":
          content:
            application/json:
              example:
                message: Сущность не найдена
                detailedMessage: Более детальное сообщение об ошибке
                dateTime: 2024-12-03T10:15:30
              schema:
                $ref: '#/components/schemas/ErrorMessageResponse'
          description: Сущность не найдена.
        "500":
          content:
            application/json:
              example:
                message: Внутренняя ошибка сервера
                detailedMessage: Более детальное сообщение об ошибке
                dateTime: 2024-12-03T10:15:30
              schema:
                $ref: '#/components/schemas/ErrorMessageResponse'
          description: Внутренняя ошибка сервера
      summary: "Обновление мероприятия. Allowed roles=[USER, ADMIN]"
      tags:
      - Events
  /events/search:
    post:
      description: Поиск всех мероприятий по фильтру
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventSearchRequestDto'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/EventDto'
                type: array
          description: "Список мероприятий, соответствующих заданным фильтрам"
        "400":
          content:
            application/json:
              example:
                message: Некорректный запрос
                detailedMessage: Более детальное сообщение об ошибке
                dateTime: 2024-12-03T10:15:30
              schema:
                $ref: '#/components/schemas/ErrorMessageResponse'
          description: "Запрос с невалидными данными. Например, попытка занять зан\
            ятый логин или некорректные параметры запроса."
        "401":
          content:
            application/json:
              example:
                message: Необходима аутентификация
                detailedMessage: Более детальное сообщение об ошибке
                dateTime: 2024-12-03T10:15:30
              schema:
                $ref: '#/components/schemas/ErrorMessageResponse'
          description: Аутентификация отсутствует или не удалась
        "403":
          content:
            application/json:
              example:
                message: Недостаточно прав для выполнения операции
                detailedMessage: Более детальное сообщение об ошибке
                dateTime: 2024-12-03T10:15:30
              schema:
                $ref: '#/components/schemas/ErrorMessageResponse'
          description: У пользователя недостаточно прав для выполнения этой операции
        "404":
          content:
            application/json:
              example:
                message: Сущность не найдена
                detailedMessage: Более детальное сообщение об ошибке
                dateTime: 2024-12-03T10:15:30
              schema:
                $ref: '#/components/schemas/ErrorMessageResponse'
          description: Сущность не найдена.
        "500":
          content:
            application/json:
              example:
                message: Внутренняя ошибка сервера
                detailedMessage: Более детальное сообщение об ошибке
                dateTime: 2024-12-03T10:15:30
              schema:
                $ref: '#/components/schemas/ErrorMessageResponse'
          description: Внутренняя ошибка сервера
      summary: "Поиск мероприятий по фильтру. Allowed roles=[USER, ADMIN]"
      tags:
      - Events
  /events/my:
    get:
      description: "Все мероприятия созданные пользователем, который выполняет зап\
        рос."
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/EventDto'
                type: array
          description: "Список мероприятий, созданных пользователем"
        "401":
          content:
            application/json:
              example:
                message: Необходима аутентификация
                detailedMessage: Более детальное сообщение об ошибке
                dateTime: 2024-12-03T10:15:30
              schema:
                $ref: '#/components/schemas/ErrorMessageResponse'
          description: Аутентификация отсутствует или не удалась
        "403":
          content:
            application/json:
              example:
                message: Недостаточно прав для выполнения операции
                detailedMessage: Более детальное сообщение об ошибке
                dateTime: 2024-12-03T10:15:30
              schema:
                $ref: '#/components/schemas/ErrorMessageResponse'
          description: У пользователя недостаточно прав для выполнения этой операции
        "500":
          content:
            application/json:
              example:
                message: Внутренняя ошибка сервера
                detailedMessage: Более детальное сообщение об ошибке
                dateTime: 2024-12-03T10:15:30
              schema:
                $ref: '#/components/schemas/ErrorMessageResponse'
          description: Внутренняя ошибка сервера
      summary: "Получение всех мероприятий, созданных текущим пользователем. Allowed\
        \ roles=[USER]"
      tags:
      - Events
  /events/registrations/{eventId}:
    post:
      description: "Необходимо учесть, что статус мероприятия должен позволять рег\
        истрацию. Можно зарегестрироваться только на мероприятие, которое не закон\
        очилось и не отменено."
      parameters:
      - description: Уникальный идентификатор мероприятия
        in: path
        name: eventId
        required: true
        schema:
          example: 10
          format: int64
          type: integer
      responses:
        "200":
          description: Успешная регистрация на мероприятие
        "400":
          content:
            application/json:
              example:
                message: Некорректный запрос
                detailedMessage: Более детальное сообщение об ошибке
                dateTime: 2024-12-03T10:15:30
              schema:
                $ref: '#/components/schemas/ErrorMessageResponse'
          description: "Запрос с невалидными данными. Например, попытка занять зан\
            ятый логин или некорректные параметры запроса."
        "401":
          content:
            application/json:
              example:
                message: Необходима аутентификация
                detailedMessage: Более детальное сообщение об ошибке
                dateTime: 2024-12-03T10:15:30
              schema:
                $ref: '#/components/schemas/ErrorMessageResponse'
          description: Аутентификация отсутствует или не удалась
        "403":
          content:
            application/json:
              example:
                message: Недостаточно прав для выполнения операции
                detailedMessage: Более детальное сообщение об ошибке
                dateTime: 2024-12-03T10:15:30
              schema:
                $ref: '#/components/schemas/ErrorMessageResponse'
          description: У пользователя недостаточно прав для выполнения этой операции
        "404":
          content:
            application/json:
              example:
                message: Сущность не найдена
                detailedMessage: Более детальное сообщение об ошибке
                dateTime: 2024-12-03T10:15:30
              schema:
                $ref: '#/components/schemas/ErrorMessageResponse'
          description: Сущность не найдена.
        "500":
          content:
            application/json:
              example:
                message: Внутренняя ошибка сервера
                detailedMessage: Более детальное сообщение об ошибке
                dateTime: 2024-12-03T10:15:30
              schema:
                $ref: '#/components/schemas/ErrorMessageResponse'
          description: Внутренняя ошибка сервера
      summary: "Регистрация пользователя на мероприятие по ID. Allowed roles=[USER]"
      tags:
      - Events
  /events/registrations/cancel/{eventId}:
    delete:
      description: "Необходимо учесть статус мероприятия. Нельзя отменить регистра\
        цию, если мероприятие уже началось или закончилось."
      parameters:
      - description: Уникальный идентификатор мероприятия
        in: path
        name: eventId
        required: true
        schema:
          example: 10
          format: int64
          type: integer
      responses:
        "204":
          description: Успешная отмена регистрации на мероприятие
        "400":
          content:
            application/json:
              example:
                message: Некорректный запрос
                detailedMessage: Более детальное сообщение об ошибке
                dateTime: 2024-12-03T10:15:30
              schema:
                $ref: '#/components/schemas/ErrorMessageResponse'
          description: "Запрос с невалидными данными. Например, попытка занять зан\
            ятый логин или некорректные параметры запроса."
        "401":
          content:
            application/json:
              example:
                message: Необходима аутентификация
                detailedMessage: Более детальное сообщение об ошибке
                dateTime: 2024-12-03T10:15:30
              schema:
                $ref: '#/components/schemas/ErrorMessageResponse'
          description: Аутентификация отсутствует или не удалась
        "403":
          content:
            application/json:
              example:
                message: Недостаточно прав для выполнения операции
                detailedMessage: Более детальное сообщение об ошибке
                dateTime: 2024-12-03T10:15:30
              schema:
                $ref: '#/components/schemas/ErrorMessageResponse'
          description: У пользователя недостаточно прав для выполнения этой операции
        "404":
          content:
            application/json:
              example:
                message: Сущность не найдена
                detailedMessage: Более детальное сообщение об ошибке
                dateTime: 2024-12-03T10:15:30
              schema:
                $ref: '#/components/schemas/ErrorMessageResponse'
          description: Сущность не найдена.
        "500":
          content:
            application/json:
              example:
                message: Внутренняя ошибка сервера
                detailedMessage: Более детальное сообщение об ошибке
                dateTime: 2024-12-03T10:15:30
              schema:
                $ref: '#/components/schemas/ErrorMessageResponse'
          description: Внутренняя ошибка сервера
      summary: "Отмена регистрации на мероприятие. Allowed roles=[USER]"
      tags:
      - Events
  /events/registrations/my:
    get:
      description: "Все мероприятия на которые записан пользователь. Мероприятия д\
        олжны возвращаться все, даже те, которые отменены или закончены."
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/EventDto'
                type: array
          description: "Список мероприятий, на которые пользователь зарегистрирова\
            н"
        "401":
          content:
            application/json:
              example:
                message: Необходима аутентификация
                detailedMessage: Более детальное сообщение об ошибке
                dateTime: 2024-12-03T10:15:30
              schema:
                $ref: '#/components/schemas/ErrorMessageResponse'
          description: Аутентификация отсутствует или не удалась
        "403":
          content:
            application/json:
              example:
                message: Недостаточно прав для выполнения операции
                detailedMessage: Более детальное сообщение об ошибке
                dateTime: 2024-12-03T10:15:30
              schema:
                $ref: '#/components/schemas/ErrorMessageResponse'
          description: У пользователя недостаточно прав для выполнения этой операции
        "500":
          content:
            application/json:
              example:
                message: Внутренняя ошибка сервера
                detailedMessage: Более детальное сообщение об ошибке
                dateTime: 2024-12-03T10:15:30
              schema:
                $ref: '#/components/schemas/ErrorMessageResponse'
          description: Внутренняя ошибка сервера
      summary: "Получение мероприятий, на которые зарегистрирован текущий пользова\
        тель. Allowed roles=[USER]"
      tags:
      - Events
components:
...
  schemas:
    ErrorMessageResponse:
      example:
        message: Сообщение об ошибке
        detailedMessage: Более детальное сообщение об ошибке
        dateTime: 2024-12-03T10:15:30
      properties:
        message:
          description: Человеко-читаемое сообщение об ошибке
          type: string
        detailedMessage:
          description: "Более детальное сообщение об ошибке. Например, содержимое\
            \ exception"
          type: string
        dateTime:
          description: Дата и время возникновения ошибки. Формат "YYYY-MM-DDThh:mm:ss"
          format: date-time
          type: string
      required:
      - dateTime
      - detailedMessage
      - message
      type: object
...
    EventDto:
      example:
        occupiedPlaces: 15
        date: 2025-01-23T04:56:07.000
        duration: 60
        cost: 1200
        maxPlaces: 10
        locationId: 0
        name: Лекция по Java
        id: 42
        ownerId: "10"
        status: WAIT_START
      properties:
        id:
          example: 42
          format: int64
          type: integer
        name:
          description: Название мероприятия
          example: Лекция по Java
          type: string
        ownerId:
          description: id пользователя-создателя мероприятия
          example: "10"
          type: string
        maxPlaces:
          description: Максимальное кол-во мест на мероприятии
          example: 10
          type: integer
        occupiedPlaces:
          description: Кол-во уже занятых мест (создатель не учитывается при подсчете)
          example: 15
          type: integer
        date:
          description: Дата и время проведения мероприятия. Формат "YYYY-MM-DDThh:mm:ss"
          format: date-time
          type: string
        cost:
          description: Стоимость в рублях
          example: 1200
          minimum: 0
          type: decimal
        duration:
          description: Длительность в минутах
          example: 60
          minimum: 30
          type: integer
        locationId:
          description: "Идентификатор локации, где проходит мероприятие"
          type: integer
        status:
          $ref: '#/components/schemas/EventStatus'
      required:
      - cost
      - date
      - duration
      - id
      - locationId
      - maxPlaces
      - name
      - occupiedPlaces
      - ownerId
      type: object
...
    EventCreateRequestDto:
      description: Объект с данными для создания нового мероприятия.
      example:
        date: 2025-01-23T04:56:07.000
        duration: 60
        cost: 1200
        maxPlaces: 10
        locationId: 10
        name: Лекция по Java
      properties:
        name:
          description: Название мероприятия
          example: Лекция по Java
          type: string
        maxPlaces:
          description: Максимальное кол-во мест на мероприятии
          example: 10
          type: integer
        date:
          description: Дата и время проведения мероприятия. Формат "YYYY-MM-DDThh:mm:ss"
          format: date-time
          type: string
        cost:
          description: Стоимость в рублях
          example: 1200
          minimum: 0
          type: integer
        duration:
          description: Длительность в минутах
          example: 60
          minimum: 30
          type: integer
        locationId:
          description: "Идентификатор локации, где проходит мероприятие"
          example: 10
          type: integer
      required:
      - cost
      - date
      - duration
      - locationId
      - maxPlaces
      - name
      type: object
    EventUpdateRequestDto:
      description: Объект с данными для обновления мероприятия
      example:
        date: 2025-01-23T04:56:07.000
        duration: 60
        cost: 3000
        maxPlaces: 100
        locationId: 50
        name: Новое назание
      properties:
        name:
          description: Новое название мероприятия
          example: Новое назание
          type: string
        maxPlaces:
          description: "Новое кол-во мест мероприятия. Локация должна позволять вм\
            естить столько людей. Должно быть больше, чем уже записанных людей."
          example: 100
          type: integer
        date:
          description: Новая дата и время проведения мероприятия. Формат "YYYY-MM-DDThh:mm:ss"
          format: date-time
          type: string
        cost:
          description: Новая стоимость мероприятия в рублях
          example: 3000
          minimum: 0
          type: decimal
        duration:
          description: Новая длительность в минутах
          example: 60
          minimum: 30
          type: integer
        locationId:
          description: "Новая локация для мероприятия. Локация должна вмещать коли\
            чество пользователей, которые уже записаны. Также локация должна вмеща\
            ть maxPlaces - кол-во мест на мероприятии."
          example: 50
          type: integer
      type: object
...
  securitySchemes:
    bearerAuth:
      bearerFormat: JWT
      scheme: bearer
      type: http